refactor: move playEmote into its own funtion and file
